requirements.txt
pipecat-ai
daily-python
python-dotenv
loguru
aiohttp
